+++ baybe/acquisition/base.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import gc
-import warnings
 from abc import ABC
 from typing import TYPE_CHECKING, ClassVar, Literal, overload
 
@@ -183,40 +182,8 @@ def _get_botorch_acqf_class(
 
 
 # Register (un-)structure hooks
-def _add_deprecation_hook(hook):
-    """Add deprecation warnings to the default hook.
-
-    Used for backward compatibility only and will be removed in future versions.
-    """
-
-    def added_deprecation_hook(val: dict | str, cls: type):
-        # Backwards-compatibility needs to be ensured only for deserialization from
-        # base class using string-based type specifiers as listed below,
-        # since the concrete classes were available only after the change.
-        if is_abstract(cls):
-            UCB_DEPRECATIONS = {
-                "VarUCB": "UpperConfidenceBound",
-                "qVarUCB": "qUpperConfidenceBound",
-            }
-            if (
-                entry := val if isinstance(val, str) else val["type"]
-            ) in UCB_DEPRECATIONS:
-                warnings.warn(
-                    f"The use of `{entry}` is deprecated and will be disabled in a "
-                    f"future version. To get the same outcome, use the new "
-                    f"`{UCB_DEPRECATIONS[entry]}` class instead with a beta of 100.0.",
-                    DeprecationWarning,
-                )
-                val = {"type": UCB_DEPRECATIONS[entry], "beta": 100.0}
-
-        return hook(val, cls)
-
-    return added_deprecation_hook
-
-
 converter.register_structure_hook(
-    AcquisitionFunction,
-    _add_deprecation_hook(get_base_structure_hook(AcquisitionFunction)),
+    AcquisitionFunction, get_base_structure_hook(AcquisitionFunction)
 )
 converter.register_unstructure_hook(AcquisitionFunction, unstructure_base)