+++ baybe/recommenders/pure/bayesian/base.py
@@ -16,7 +16,6 @@ from baybe.acquisition import qLogEI, qLogNEHVI
 from baybe.acquisition.base import AcquisitionFunction
 from baybe.acquisition.utils import convert_acqf
 from baybe.exceptions import (
-    DeprecationError,
     IncompatibleAcquisitionFunctionError,
     InvalidSurrogateModelError,
 )
@@ -68,18 +67,6 @@ class BayesianRecommender(PureRecommender, ABC):
 
     _botorch_acqf = field(default=None, init=False, eq=False)
     """The induced BoTorch acquisition function."""
-
-    acquisition_function_cls: str | None = field(default=None, kw_only=True)
-    "Deprecated! Raises an error when used."
-
-    @acquisition_function_cls.validator
-    def _validate_deprecated_argument(self, _, value) -> None:
-        """Raise DeprecationError if old acquisition_function_cls parameter is used."""
-        if value is not None:
-            raise DeprecationError(
-                "Passing 'acquisition_function_cls' to the constructor is deprecated. "
-                "The parameter has been renamed to 'acquisition_function'."
-            )
 
     @property
     def surrogate_model(self) -> SurrogateProtocol: