+++ CHANGELOG.md
@@ -74,6 +74,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `fuzzy_row_match` will no longer warn about entries not matching to the search space 
 - `funcy` dependency
 
+### Expired Deprecations (from 0.9.*)
+- `baybe.objective` namespace 
+- `acquisition_function_cls` constructor parameter for `BayesianRecommender`
+- `VarUCB` and `qVarUCB` acquisition functions
+
 ## [0.12.2] - 2025-01-31
 ### Changed
 - More robust settings for the GP fitting